+++ app/models/manageiq/providers/ansible_tower/automation_manager.rb
@@ -1,20 +1,25 @@
 class ManageIQ::Providers::AnsibleTower::AutomationManager < ManageIQ::Providers::ExternalAutomationManager
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager
+
+  require_nested :AmazonCredential
+  require_nested :AzureCredential
+  require_nested :CloudCredential
+  require_nested :GoogleCredential
+  require_nested :MachineCredential
+  require_nested :NetworkCredential
+  require_nested :OpenstackCredential
+  require_nested :RackspaceCredential
+  require_nested :Satellite6Credential
+  require_nested :VmwareCredential
+
   require_nested :ConfigurationScript
   require_nested :ConfiguredSystem
+  require_nested :EventCatcher
+  require_nested :EventParser
+  require_nested :Job
+  require_nested :Playbook
   require_nested :Refresher
   require_nested :RefreshWorker
-  require_nested :Job
-
-  include ProcessTasksMixin
-  delegate :authentications,
-           :authentication_check,
-           :authentication_status,
-           :authentication_status_ok?,
-           :connect,
-           :missing_credentials?,
-           :verify_credentials,
-           :with_provider_connection,
-           :to => :provider
 
   def self.ems_type
     @ems_type ||= "ansible_tower_automation".freeze
@@ -22,15 +27,5 @@ class ManageIQ::Providers::AnsibleTower::AutomationManager < ManageIQ::Providers
 
   def self.description
     @description ||= "Ansible Tower Automation".freeze
-  end
-
-  def image_name
-    "ansible_tower_automation"
-  end
-
-  private
-
-  def connection_source(options = {})
-    options[:connection_source] || self
   end
 end
