+++ app/models/manageiq/providers/ansible_tower/automation_manager/azure_credential.rb
@@ -1,3 +1,4 @@
-# This corresponds to Ansible Tower's Azure Resource Manager (azure_rm) type credential.  We are not modeling the deprecated Azzure classic
-class ManageIQ::Providers::AnsibleTower::AutomationManager::AzureCredential < ManageIQ::Providers::AnsibleTower::AutomationManager::CloudCredential
+# This corresponds to Ansible Tower's Azure Resource Manager (azure_rm) type credential.  We are not modeling the deprecated Azure classic
+class ManageIQ::Providers::AnsibleTower::AutomationManager::AzureCredential <
+  ManageIQ::Providers::AnsibleTower::AutomationManager::CloudCredential
 end
