+++ app/models/manageiq/providers/ansible_tower/automation_manager/configured_system.rb
@@ -1,17 +1,6 @@
-class ManageIQ::Providers::AnsibleTower::AutomationManager::ConfiguredSystem < ::ConfiguredSystem
+class ManageIQ::Providers::AnsibleTower::AutomationManager::ConfiguredSystem <
+  ManageIQ::Providers::ExternalAutomationManager::ConfiguredSystem
+
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::ConfiguredSystem
   include ProviderObjectMixin
-
-  def provider_object(connection = nil)
-    (connection || connection_source.connect).api.hosts.find(manager_ref)
-  end
-
-  def ext_management_system
-    manager
-  end
-
-  private
-
-  def connection_source(options = {})
-    options[:connection_source] || manager
-  end
 end
