+++ app/models/manageiq/providers/ansible_tower/automation_manager/credential.rb
@@ -1,2 +1,2 @@
-class ManageIQ::Providers::AnsibleTower::AutomationManager::Credential < ManageIQ::Providers::AutomationManager::Authentication
+class ManageIQ::Providers::AnsibleTower::AutomationManager::Credential < ManageIQ::Providers::ExternalAutomationManager::Authentication
 end
