+++ app/models/manageiq/providers/ansible_tower/automation_manager/playbook.rb
@@ -1,2 +1,3 @@
-class ManageIQ::Providers::AnsibleTower::AutomationManager::Playbook < ConfigurationScriptPayload
+class ManageIQ::Providers::AnsibleTower::AutomationManager::Playbook <
+  ManageIQ::Providers::ExternalAutomationManager::ConfigurationScriptPayload
 end
