+++ app/models/manageiq/providers/ansible_tower/automation_manager/scm_credential.rb
@@ -1,2 +1,2 @@
-class ManageIQ::Providers::AnsibleTower::AutomationManager::ScmCredential < ManageIQ::Providers::AutomationManager::Authentication
+class ManageIQ::Providers::AnsibleTower::AutomationManager::ScmCredential < ManageIQ::Providers::AnsibleTower::AutomationManager::Credential
 end
