+++ app/models/manageiq/providers/automation_manager.rb
@@ -1,7 +1,11 @@
 class ManageIQ::Providers::AutomationManager < ::ExtManagementSystem
+  require_nested :Authentication
+  require_nested :ConfigurationScript
+  require_nested :ConfigurationScriptPayload
+  require_nested :ConfiguredSystem
   require_nested :InventoryGroup
   require_nested :InventoryRootGroup
-  require_nested :Authentication
+  require_nested :OrchestrationStack
 
   has_many :configured_systems,           :dependent => :destroy, :foreign_key => "manager_id"
   has_many :configuration_profiles,       :dependent => :destroy, :foreign_key => "manager_id"
