+++ app/models/manageiq/providers/automation_manager/configuration_script.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::AutomationManager::ConfigurationScript < ::ConfigurationScript
+end
