+++ app/models/manageiq/providers/automation_manager/configuration_script_payload.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::AutomationManager::ConfigurationScriptPayload < ::ConfigurationScriptPayload
+end
