+++ app/models/manageiq/providers/automation_manager/configured_system.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::AutomationManager::ConfiguredSystem < ::ConfiguredSystem
+end
