+++ app/models/manageiq/providers/automation_manager/orchestration_stack.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::AutomationManager::OrchestrationStack < ::OrchestrationStack
+end
