+++ app/models/manageiq/providers/embedded_ansible/automation_manager/amazon_credential.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::AmazonCredential < ManageIQ::Providers::EmbeddedAnsible::AutomationManager::CloudCredential
+end
