+++ app/models/manageiq/providers/embedded_ansible/automation_manager/cloud_credential.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::CloudCredential <
+  ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Credential
+end
