+++ app/models/manageiq/providers/embedded_ansible/automation_manager/configuration_script.rb
@@ -1,0 +1,5 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::ConfigurationScript <
+  ManageIQ::Providers::EmbeddedAutomationManager::ConfigurationScript
+
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::ConfigurationScript
+end
