+++ app/models/manageiq/providers/embedded_ansible/automation_manager/configured_system.rb
@@ -1,0 +1,6 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::ConfiguredSystem <
+  ManageIQ::Providers::EmbeddedAutomationManager::ConfiguredSystem
+
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::ConfiguredSystem
+  include ProviderObjectMixin
+end
