+++ app/models/manageiq/providers/embedded_ansible/automation_manager/credential.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Credential <
+  ManageIQ::Providers::EmbeddedAutomationManager::Authentication
+end
