+++ app/models/manageiq/providers/embedded_ansible/automation_manager/job.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Job < ManageIQ::Providers::EmbeddedAutomationManager::OrchestrationStack
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::Job
+end
