+++ app/models/manageiq/providers/embedded_ansible/automation_manager/job/status.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Job::Status < ::OrchestrationStack::Status
+  include ManageIQ::Providers::AnsibleTower::Shared::AutomationManager::Job::Status
+end
