+++ app/models/manageiq/providers/embedded_ansible/automation_manager/openstack_credential.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::OpenstackCredential < ManageIQ::Providers::EmbeddedAnsible::AutomationManager::CloudCredential
+end
