+++ app/models/manageiq/providers/embedded_ansible/automation_manager/playbook.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Playbook <
+  ManageIQ::Providers::EmbeddedAutomationManager::ConfigurationScriptPayload
+end
