+++ app/models/manageiq/providers/embedded_ansible/automation_manager/satellite6_credential.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Satellite6Credential < ManageIQ::Providers::EmbeddedAnsible::AutomationManager::CloudCredential
+end
