+++ app/models/manageiq/providers/embedded_ansible/automation_manager/scm_credential.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::ScmCredential < ManageIQ::Providers::EmbeddedAnsible::AutomationManager::Credential
+end
