+++ app/models/manageiq/providers/embedded_ansible/automation_manager/vmware_credential.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::EmbeddedAnsible::AutomationManager::VmwareCredential < ManageIQ::Providers::EmbeddedAnsible::AutomationManager::CloudCredential
+end
