+++ app/models/manageiq/providers/embedded_automation_manager/authentication.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAutomationManager::Authentication <
+  ManageIQ::Providers::AutomationManager::Authentication
+end
