+++ app/models/manageiq/providers/embedded_automation_manager/configuration_script.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAutomationManager::ConfigurationScript <
+  ManageIQ::Providers::AutomationManager::ConfigurationScript
+end
