+++ app/models/manageiq/providers/embedded_automation_manager/configuration_script_payload.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAutomationManager::ConfigurationScriptPayload <
+  ManageIQ::Providers::AutomationManager::ConfigurationScriptPayload
+end
