+++ app/models/manageiq/providers/embedded_automation_manager/configured_system.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::EmbeddedAutomationManager::ConfiguredSystem < ManageIQ::Providers::AutomationManager::ConfiguredSystem
+end
