+++ app/models/manageiq/providers/embedded_automation_manager/orchestration_stack.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::EmbeddedAutomationManager::OrchestrationStack <
+  ManageIQ::Providers::AutomationManager::OrchestrationStack
+end
