+++ app/models/manageiq/providers/external_automation_manager.rb
@@ -1,2 +1,7 @@
 class ManageIQ::Providers::ExternalAutomationManager < ManageIQ::Providers::AutomationManager
+  require_nested :Authentication
+  require_nested :ConfigurationScript
+  require_nested :ConfigurationScriptPayload
+  require_nested :ConfiguredSystem
+  require_nested :OrchestrationStack
 end
