+++ app/models/manageiq/providers/external_automation_manager/authentication.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::ExternalAutomationManager::Authentication <
+  ManageIQ::Providers::AutomationManager::Authentication
+end
