+++ app/models/manageiq/providers/external_automation_manager/configuration_script.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::ExternalAutomationManager::ConfigurationScript <
+  ManageIQ::Providers::AutomationManager::ConfigurationScript
+end
