+++ app/models/manageiq/providers/external_automation_manager/configuration_script_payload.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::ExternalAutomationManager::ConfigurationScriptPayload <
+  ManageIQ::Providers::AutomationManager::ConfigurationScriptPayload
+end
