+++ app/models/manageiq/providers/external_automation_manager/configured_system.rb
@@ -1,0 +1,2 @@
+class ManageIQ::Providers::ExternalAutomationManager::ConfiguredSystem < ManageIQ::Providers::AutomationManager::ConfiguredSystem
+end
