+++ app/models/manageiq/providers/external_automation_manager/orchestration_stack.rb
@@ -1,0 +1,3 @@
+class ManageIQ::Providers::ExternalAutomationManager::OrchestrationStack <
+  ManageIQ::Providers::AutomationManager::OrchestrationStack
+end
